ENH: SlicerQT - Added qVTKCommandOptions allowing to handle command line parameter in a standard fashion.
Note that this class is made to be subclassed. That way the method:
  intialize, disableCurrentSettings, wrongArgument, deprecatedArgument
could be overloaded

An interesting aspect is that the class is also working with QSettings,

For example:
  In settings file, there is the following property:  tmpDir=/tmp

class MyAppCommandOptions: public qVTKCommandOptions
{
public:
  QString TmpDir;

  QString tmpDir();
 void initialize()
  {
  Superclass::initialize();
  this->addArgument("--temp-dir", 0, &TmpDir, QDir::tmp())
  }

  ...
}

bash$ bin/SlicerQT --temp-dir=/home/mytmp

At run-time the application will be using the specified directory. In other
word the function .. options->tmpDir() will return "/home/mytmp"

Then, next time the application is restarted .. if no option is given,
the tmpDir() method will return the value stored in the setting file.

The following two parameters are available by default: --help and --disable-settings
--help will show the different available arguments

--disable-settings will call the virtual method 'disableCurrentSettings'.
The derived class a responsible to disable the settings. A possible way
could be to instanciate a settings object using a filename + '-disabled' suffix.
That way ... no parameter will be used.

git-svn-id: http://svn.slicer.org/Slicer3/trunk@11984 3bd1e089-480b-0410-8dfb-8563597acbee