BUG: Fix relative path of save nodes was incorrect + code refactoring
Depending on what was saved first (scene or nodes) the paths were not
correctly set. To save the mrml scene, the new paths of the nodes need
to be set. To save the mrml nodes, the new path of the scene needs to
be set (so that the mrml scene use the correct saved mrml node).

Factoring of the code.
Remove trailing whitespaces

git-svn-id: http://svn.slicer.org/Slicer3/trunk@13776 3bd1e089-480b-0410-8dfb-8563597acbee