ENH: SlicerLayoutManager is now handling properly when scene are imported or loaded
Internally, the ThreeDRenderView are not anymore stored in a map
using a string of the form "RenderView#ID".

There are now two lists: ThreeDRenderViewList and MRMLViewNodeList
respectively storing the reference to the ThreeDRenderView and its associated vtkMRMLViewNode.

It means the ThreeDRenderView and associated node are identified with an integer
corresponding to:
 0 - If first ViewNode added
 1 - If second ViewNode added
 2 - ...

A reference to an already instantiated ThreeDRenderView could be "obtained"
using either an integer or vtkMRMLViewNode pointer.


The SliceViewWidget are still identified using the LayoutName associated with the SliceViewNode.

When a scene is loaded (using ::Connect), the current ViewArrangement is saved.
That allows, in case no layoutNode were loaded from the file, to restore the view
the user was using.

git-svn-id: http://svn.slicer.org/Slicer3/trunk@14172 3bd1e089-480b-0410-8dfb-8563597acbee