ENH: vtkSliceLogic and vtkMRMLAbstractLogic now provide a single way of initialization: setMRMLSceneInternal
Before that changes, it was the role of the client (user of the class) to call
either SetMRMLScene, SetAndObserveMRMLScene or SetAndObserveMRMLSceneEvent.

The way to understand what was the correct way of initialization of the logic
was to find some code already using the logic and copy/paste the proper statements.

There is now one virtual method names: SetMRMLSceneInternal called within
SetMRMLScene.

The call to SetMRMLSceneInternal is done just before
RegisterNodes and invoke(ModifiedEvent).

If specific initialization are required, such as observing the scene for
specific event.

The method SetMRMLSceneInternal should be overloaded.
For example, if a module should listen for NodeAddedEvent and NodeRemovedEvent
, it could overload the method like stated below:

virtual void SetMRMLSceneInternal(vtkMRMLScene* newScene)
{
  // List of events the slice logics should listen
  vtkSmartPointer<vtkIntArray> events = vtkSmartPointer<vtkIntArray>::New();
  events->InsertNextValue(vtkMRMLScene::NodeAddedEvent);
  events->InsertNextValue(vtkMRMLScene::NodeRemovedEvent);

  this->SetAndObserveMRMLSceneEventsInternal(newScene, events);
}

The method SetAndObserveMRMLSceneEventsInternal and SetAndObserveMRMLSceneInternal
are provided for conveniency and provide an easy way to interface with the vtkObserverManager

Existing methods SetAndObserveMRMLScene and SetAndObserveMRMLSceneEvents
are now deprecated.

All modules ported to Qt should overload SetMRMLSceneInternal and provide
a single way of initialization.


Future...

In case the user of class desire to change the list of observed events,
the required logic should be added in the base classes: vtkSlicerLogic and
vtkMRMLAbstractLogic

The user of the class should be able to query the object and now which event
it is listening and it could optionally configured not to listen some event.
Note also that such options should be tested and thoroughly documented.


Remark:
This commit also fix issue related to qSlicerSliceControllerModuleWidget.
Note also that the controller widget added in the "slices controller" module are
configured with the same logic used in the main slice viewer/controller.
For that reason, the SliceLogic has a method Initialize/IsInitialized avoiding
additional overhead and creation of extra vtkMRMLSliceNode.

git-svn-id: http://svn.slicer.org/Slicer3/trunk@14352 3bd1e089-480b-0410-8dfb-8563597acbee