ENH: Updated qMRMLSliceViewWidget to consider vtkMRMLSliceViewDisplayableManagerFactory
Note that since SliceView deals internally with a lightBox (a grid of renderer)
for now the first renderer of the grid is passed.

It could also be possible to instantiate as many DisplayableManagerGroup
as there are Renderer in the SliceView (one for each item of the grid).
Doing so would require to add signal to ctkVTKSliceView allowing to
know when the Layout is changed in order to instantiate
or delete displatableManagerGroup as needed.

git-svn-id: http://svn.slicer.org/Slicer3/trunk@14366 3bd1e089-480b-0410-8dfb-8563597acbee