BUG: When "Adding a Volume", the layout was not restored in SlicerQT
"Adding a Volume" doesn't close the scene, the SavedLayout was then not
set to the current layout but containing None (value by default).
After the volume is imported, the wrong layout (none) was restored.
The fix is to set the layout based on the value of the layout node instead.

git-svn-id: http://svn.slicer.org/Slicer3/trunk@14480 3bd1e089-480b-0410-8dfb-8563597acbee